requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
transformers>=4.30.0
fastapi>=0.100.0
uvicorn>=0.23.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
jiwer>=3.0.0
datasets>=2.14.0
pyarrow>=13.0.0
sam-2>=1.0.0
